Trait pwasm_abi::eth::AbiType

pub trait AbiType: Sized {
    const IS_FIXED: bool;

    fn decode(stream: &mut Stream) -> Result<Self, Error>;

    fn encode(self, sink: &mut Sink);
}

Abi type trait

Associated Constants

const IS_FIXED: bool

Whether type has fixed length or not

Required Methods

fn decode(stream: &mut Stream) -> Result<Self, Error>

Insantiate type from data stream Should never be called manually! Use stream.pop()

fn encode(self, sink: &mut Sink)

Push type to data sink Should never be called manually! Use sink.push(val)