Trait pwasm_abi::eth::AsLog

pub trait AsLog {
    fn as_log(&self) -> H256;
}

As log trait for how primitive types are represented as indexed arguments of the event log

Required Methods

fn as_log(&self) -> H256

Convert type to hash representation for the event log.