Struct pwasm_abi::eth::Sink

pub struct Sink { /* fields omitted */ }

Sink for returning number of arguments

Methods

impl Sink

pub fn new(capacity: usize) -> Self

New sink with known capacity

pub fn push<T: AbiType>(&mut self, val: T)

Consume val to the Sink

pub fn drain_to(self, target: &mut Vec<u8>)

Drain current Sink to the target vector

pub fn finalize_panicking(self) -> Vec<u8>

Consume current Sink to produce a vector with content. May panic if declared number of arguments does not match the resulting number of bytes should be produced.

pub fn preamble_mut(&mut self) -> &mut Vec<u8>

Mutable reference to the Sink preamble

pub fn heap_mut(&mut self) -> &mut Vec<u8>

Mutable reference to the Sink heap